Rename tools to reminderadd, reminderlist and reminderremove.
