+++ packages/opencode/src/tool/reminderremove.txt
@@ -1,0 +1,1 @@
+Cancel a scheduled reminder. Use when user asks to 'stop checking X' or 'cancel the reminder for Y'. Will attempt to match user's description to existing reminders.
